GH #1321: Use pkl file to access lick information in ophys_sessions Sometimes the sync file does not contain lick data. This PR modifies the source of lick information to be the "behavior pickle" file. Note that this change is not backwards compatible. Though the files are mostly equivalent, the "behavior pickle" file is less temporally precise. This means that re-running old analyses may produce slightly different results. The decision was made to make this change to include the experiments that do not have sync file licks in the data release.
Fix time stream alignment for licks in BehaviorSessions.
The licks from the pickle file used the information coded
in the `intervalsms` time series, which was not properly
aligned to the `trial_log` time stream. We now calculate
the time to the first `intervalsms` with a linear
regression of frame and time from the `trial_log` time
stream, and add the time offset to the `intervalsms`
time series.