Added version for AIBS_ophys_behavior_namespace
NWB allows you to create extensions composed of custom data structures.
Allen has been using this feature to add meta-data that is not in the core NWB.
As we find better ways to store this data, and as Allen's needs evolve,
we will want to change the extensions to meet these needs. This version number
is the version of the extension AIBS_ophys_behavior_extension.yaml. It allows us to make
changes to the extension in an organized way so we can have an understanding
of what state of the extension was used for a certain file even as we update it for other files.

Resolves: #1359 #1382