GH #900: Compute display lag
Previously the BehaviorOphysLimsApi did not compute
monitor delay, but added a default value of 0.351.
This updates the default to 0.215, which is the
average of monitor delay for a large number of recent
datasets.

Additionally use the OphysTimeAligner methods to
compute the display lag, which is used in the
OphysTimeSync job to produce corrected timestamps
in the LIMS pipeline.