GH-1404: Handle monitor delay errors
If there are more "stimulus transitions" than "photodiode events",
truncate the "stimulus transitions" to the number of
"photodiode events". This is similar to what is done with
ophys timestamps data in this module. Update test coverage
to handle this case.

Update logging to show the min and max delay, to help
with debugging in the case of unusual/default values.