Updated the get_keys method to search for correct keys
The get_keys method in time_sync.py was providing the wrong line labels
to the classes using the method. It was naively assuming that keys
existed in one set or the other with a clean division. This was not the
case and the issue was unveiled when error handling was removed from
monitor delay in #1438.

The method now searches for the possible key values and if they do not
exist assumes an old file where not all keys are present. In the case
of key not present it defaults to using the oldest key denoted in
VERSION_1_KEYS. This fixes the failing bamboo tests and is a more
robust search for the future.

Resolves: #1513
See also: #1438, #1519, #1451, #1507