GH-1363: Add 'date_of_acquisition' to behavior session summary table
'date_of_acquisition' field was added to behavior_sessions
table in the LIMS database. It is not backfilled, so there
will be nulls for all experiments done before the end of
May. Also adds the 'age_in_days' field to the behavior_session
table (required 'date_of_acquisition' to compute the age
of the mouse).