GH-1324: Update behavior sessions time sources
See http://confluence.corp.alleninstitute.org/pages/viewpage.action?pageId=80277925&moved=true
for complete documentation of the different time sources
in the data and how they are used by the APIs that populate
BehaviorDataSession and BehaviorOphysSession attributes
from LIMS data.

This PR aligns all the BehaviorDataSession timestamps to use
what's described in the linked document as 'trial_pkl_frame_times'.
These are the cumulative sum of the offset times between frames
aligned to the (frame, time) data points in the trial event log.
That alignment is done by using a linear regression to compute the
time that elapsed before frame 0 was presented, and then adding it
to the cumulative sum of the offset times.