Added 2019 production stimulus files to Stimulus Processing file list
Added the 2019 production stimulus file paths so that the behavior sessions associated with these files are unblocked. These stimulus files are currently preventing sessions from having analyzed as the stimuli cannot be loaded. This is a quick change that will get the situation unblocked but puprosefully does not present a long term fix as that work has yet to be prioritized and scoped.