Update nwb get_stimulus_presentations
This commit fixes how NaN values are handled when
loading `stimulus_presentations` from updated ecephys nwb
files. The 'expected' behavior is that if a column has mixed
dtypes and `str` is one of those types then all NaN values
should be replaced with an empty string ("") and all non-NaN
values should be cast as `str` type.