Added existance test for stimulus table stimulus name keys
Keys for the name of stimulus are different between eceephys and behavior ophys.
This adds in a small function to get the key represented in each data type and
raise an exception if one key is not found. This was added to allow ophys
behavior experiments to write to nwb files.