Omitted stimuli bug (#1623)
* Added function to set the stop_time for omitted stimuli for writing NWB

This function had to be added as the omitted stimuli did not by design
have an ending timestamp. These stimuli therfore could not be written to
NWB files as the interval data object in PyNWB requires an ending time.
This commit adds a function that adds a stop_time to a omitted row that
is based off the known time that omitted stimuli are shown. This allows
for omitted stimuli types to be written to NWB intervals.

* Added doc strings to stimulus_presentations.py

I added documentation for the functions in stimulus_presentations.py that I have touched during investigation. These functions are get_stimulus_presentations and get_visual_stimuli_df. I added doc strings to better explain what is happening in the code.