Added column name exceptions from column creation
Behavior Ophys NWB files were erroring because a column of name
'image_mask' already existed in the PlaneSegmentation NWB NWB object.
These columns are initilized by the PlaneSegmentation object on creation
and do not need to be added during writing. These column names were
added to the list of column exceptions as they already exist. This
allows NWB files to be written for Behavior Ophys experiments.