Updating api specific tests to use mock credentials
On my windows machine these tests fail as a result of the Postgresquery mixin and being unable to grab env variables for databases. I added in a DbCredentials namedtuple object to the failing tests to give false credentials so that the constructor no longer fails and tests run as expected.