GH #518: Fix operator for boolean matrix
When we migrated from the internal demixing module to
the brain_observatory demixing module (which were
duplicated code), a small update to the internal code
was missed. The plotting function returned a TypeError
because we were attempting to use '-=' on a boolean
numpy array. Using the bitwise 'xor' operator '^' solves
this issue. This commit re-implements that update in the
brain_observatory demixing module.

The rest of the internal code was checked to ensure that
no other changes were missed, and none were found."