GH #1323: Bugfix - Response time logic
For a given trial, the `response_time` now equals the
time of the first lick event for non-'aborted' trials.
If the trial is 'aborted' (lick occurs prior to the
stimulus change), or no licks occurred in the trial,
'response_time' is 'NaN' instead.

This differs from the previous implementation,
where 'response_time' was extracted from 'hit'
or 'false_alarm' events in the trial event log,
(which were triggered by licks) and null for
all other trial types.

Test trials_processing assertions

Update changelog for 2.2 version