Fixed error where key is assumed and gratings ignored
A bug existed within get_visual_stimuli_df where gratings were
not presumed to have any draw logs during an assertion all stimuli
had been gathered. This also assumed the 'image' key existed in the
stimuli pkl file. A change was added to not assume the 'image' keys
existance and to include the count of gratings in the asserted stimuli