Added grating metadata extraction to get_stimulus_metadata
Grating metadata now exists for all possible grating names
in the stimulus_metadata dataframe. Phase and correct_frequency are
available for the possible orientations, orientation data is saved as
part of the name of a grating. This commit also adds conditionals to
only check for image or grating metadata if the key exists in the
stimulus pkl file.