GH #1461: Enhancement - demixing modified to load movie stack in blocks
`demixer.demix_time_dep_masks()` now has a `block_size` parameter
with a default value of 1000. This allows frames from the movie
`stack` to be loaded and reshaped in blocks, thus avoiding
out-of-memory errors for very large stacks (e.g., 50 GB). `stack`
can still be passed as an array as before, but now also as an hdf5
dataset. `block_size` can be set to -1 to load full stack in one block.
`test_demixer.py` updated to test 2 `block_size` values (1000 and -1).