Fixes bug associated with gratings names
There was a bug that existed where orientation in located
in one table was an float and orientation in the set_log was a int.
This resulted in a failure when merging metadata to the stimulus
presentations as the names ended up being different when each value
was cast to a string. This has been fixed as orientation in the
presentations table will always be a float so casting all metadata names
to float before writting to string fixes this bug. A regressive test
case has been added to the test_get_stimulus_metadata test function.