Update get_current_source_density method for ecephys
pynwb 1.4.0 fixed a bug where if `__getitem__` was called on
a `MultiContainerInterface` such as a `ProcessingModule` with
only one item then any key could be used with square bracket
notation and the single item would be returned.

The fixed bug in pynwb 1.4.0 revealed that CSD had been accessed
using the incorrect key (Should be "ecephys_csd" instead of
"current_source_density").

This commit fixes this issue.