Fix internal Bamboo CI test failures
This commit aims to fix errors occurring with our
internal Bamboo CI infrastructure. The root cause
appears to be an extremely out of date base docker image
which causing problems when trying to install/test AllenSDK.

The new dockerfile spec also generates a drastically smaller
image (~6.5 GB -> ~1.5 GB).

Because the new docker image is very minimal it was discovered
that a particular internal test has a dependency on scikit-learn
which is why it has been added as a test-requirement.