Add gratings to stimulus metadata and removes 'image' key breaking dependencies in stimulus_processing.py (#1678)
* GH #1666, Adds support for gratings stimuli in stimulus_processing and concurrency in naming

Due to an assumption that all stimuli have 'images' keys stimulus_processing module was breaking
when a grating stimuli was loaded. This adds support for grating stimuli in the form of adding 
required data to stimulus_processing to be able to recreate gratings and eliminates image key 
dependency. This also adds test coverage for stimulus_processing that was sorely lacking. Lastly 
trials_processing names of grating stimuli were changed to be concurrent to the stimulus_processing format.
Resolves: #1666 
See Also: #1667