GH #1741 - running speed main bugfix
Fixes 2 errors in allensdk.brain_observatory.extract_running_speed.__main__.py
in `extract_running_speeds()`. In the if/else statement (lines 61-65),
1) the second consecutive `=` (a typo) was corrected to `\` so that `dx_rad` is
divided by median duration. 2) The if/else clauses were switched, so that the if
`use_median_duration` statement is the one that uses np.median(durations) and not v.v.