GH 1753: Don't truncate mesoscope acquisition frames
Scientifica rigs produce extra acquisition frame pulses
('ophys_frames'), so the number of ophys acquisition frames
are longer than the number of frames in the movie/trace.
A data fix was applied to truncate the length of the acquisition
frames to the length of number of frames in the movie/trace,
to remove these additional sentinel frames (#602).
This caused a problem with mesoscope data. Since the acquisition
frames are interleaved across multiple plane groups, the
length of the acquisition frames is intended to be longer than any
individual plane's movie/trace. Each plane in a given plane group
is acquired simultaneously, and the scope moves between acquiring
plane groups.
We have confirmed with MPE that mesoscope data should not
produce any additional sentinel frames. Once the mesoscope data
is split by plane group, the number of acquisition frames should
align to the number of frames in the movie/trace. This update removes
the data truncation from data produced by mesoscope rigs. If there
is a length mismatch between the split acquisition frames and the
movie/trace frames, raise an error.