Decorate abstract base class methods with abc.abstractmethod
Previously methods in the BehaviorOphysRawDataBase and
BehaviorRawDataBase only raised a NotImplementedError().
This resulted in an error only being thrown when the method
in questions was called without an override.

This commit adds the abstractmethod decorator which will result in
an error being thrown when trying to instantiate without a method
override.