Remove get_experiment_date() method from Behavior Extractors
Previously the BehaviorDataExtractorBase and BehaviorJsonExtractor
contained a get_experiment_date() method to obtain a
'date_of_acquisition' for a behavior or behavior + ophys session.

Turns out that this date can be acquired from the behavior
stimulus pickle file and has the added benefit of having a
more accurate start time.