Fix nwb-end-to-end test failure
An nwb-end-to-end test was failing for the
BehaviorOphysSession because an `eye_gaze_mapping_file_path`
property was not being populated.

Because the BehaviorOphysSession should never be referencing
any experimental data as a filepath in any format other than
NWB, the solution was to remove the `eye_gaze_mapping_file_path`
property and instead provide the BehaviorOphysNwbApi save
method for eye tracking with the method call from an
extractor class.