Fix how datetime is acquired for visual behavior sessions
This commit changes the default behavior for finding the
acquisition date for Behavior and BehaviorOphys visual behavior
sessions.

Previously, the date of acquisition was obtained from the
behavior session pkl file associated with the Behavior or BehaviorOphys
session. Now, this information will instead be obtained from
extractor classes (e.g. BehaviorLimsExtractor/BehaviorJsonExtractor).

If there is a large discrepency between the new method for
obtaining date of acquisition and the old method, a warning
will be raised.