Add Visual Behavior Session API consistency checks
Because scientists use the a LIMS query API when QC'ing data
and production uses a LIMS JSON API when writing NWB files,
it is critical to check that the BehaviorSession and
BehaviorOphysSession objects created using different APIs
end up having the same data.

This commit adds API consistency checks to write NWB modules
for the Visual Behavior project.

Relates to: #1886