Fix discrepancy in experiment_date for Visual Behavior JSON extractor
It turns out the JSON file written by LIMS writes the
acquisition_datetime (used by get_experiment_date() method)
in the local timezone. But when getting the field directly from LIMS
the datetime is already in UTC.

The other discovery is that *.pkl experiment start times are already
in UTC so there is no need to localize in the BehaviorDataTransforms
get_experiment_date() method.

This commit fixes these discrepancies.