Improve utility classes and functions for brain_observatory
This commit adds the ability to specify whether a LazyProperty
is `settable` or not. Previously LazyProperties could only be
`get` but not `set`.

This commit also fixes up the Session comparison utility function
to be more informative in terms of what is being compared when
comparing a BehaviorSession or BehaviorOphysSession.

Errant thought: May make sense to back this session comparison ability
into the Session classes themselves via __eq__?