Modify BehaviorDataTransforms.get_stimulus_presentations() output
This commit slightly modifies the DataFrame that is returned by the
get_stimulus_presentations() method so that if any columns contain
all NaN data, then they will be dropped. This is done in order to
align the output with how the stimulus presentations table is stored
in NWB (all NaN columns also dropped).