Save lick frame times column to NWB
GH-1904 made a modification to the output of the get_licks()
methods for BehaviorDataTransforms and BehaviorOphysDataTransforms.
The new output licks DataFrame has a new 'frame' column that is
critical for aligning timestamps. This new 'frame' column is
derived from data in the behavior stimulus *.pkl file that
the NWB file has no knowledge or access to.

This commit adds the 'frame' column to NWB files for lick data.