Add ability to skip adding eye tracking data for BehaviorOphysSession
Because some sessions have bad eye tracking data, it is desirable to
not write any of it to NWB.

This commit adds the ability to skip processing any eye tracking data
(and rig geometry data) associated with a BehaviorOphysSession.
This results in an NWB file that lacks any eye tracking information.

Relates to: #1942