Fix bug when comparing BehaviorSession from JSON vs NWB
The BehaviorMetadata class ands up determining `indicator`
(e.g. 'GCaMP') for both BehaviorSession and OphysExperiments.

Unfortunately, the "indicator" field does not get saved to
NWB for BehaviorSession's and gets saved to an NWB
ImagingPlane object for OphysExperiments.

Luckily, "indicator" can be determined from the "reporter_line"
field (which exists for both BehaviorSessions and OphysExperiments).

This commit adds the "indicator" field to the metadata dict
when loading BehaviorSession metadata from a behavior NWB.