Factor BasicLocalCache class out of CloudCacheBase
This commit refactors the CloudCacheBase class and splits very
basic caching methods and properties into a new BasicLocalCache
class. The BasicLocalCache is intended to be a bare bones
means of accessing local data/metadata files paths given a
project manifests and does not support fancier features like:

1. Supporting multiple dataset versions in cache
2. Keeping track of last loaded manifest
3. Constructing symlinks for valid data from previous dataset version
4. Ability to download data/metadata from S3
5. Warning of outdated manifests
6. and more!

This new BasicLocalCache is perfect for supporting the use case where
the cache is only expected to use the (and only the) latest dataset version
and where all data files are considered "local".