Add StaticLocalCache; Update behavior_project cloud_api and cache classes
This commit adds a new StaticLocalCache class which inherits from the
BasicLocalCache. It also updates the BehaviorProjectCloudApi and
VisualBehaviorOphysProjectCache initialization methods so that it is
possible to use the new StaticLocalCache.

The purpose of this commit is to add the ability to instantiate a
VisualBehaviorOphysProjectCache instance from an S3 bucket that has
been mounted as a "local" file system.

For end-users this will be as simple as:

1. Mounting the S3 bucket to desired location
2. Calling VisualBehaviorOphysProjectCache.from_local_cache() method
   and providng the mounted S3 bucket location as the "cache_dir" and
   setting the "use_static_cache" parameter to True