remove from_lims, to_json and from_json from RunningSpeed/Acquisition
Now that are are going to be able to instantiate these
objects from different combinations of stimulus and sync files,
there is no simple way to instantiate them from a single
dict_repr or LIMS database query.

This commit also makes sure that the timestamps associated with
the running objects do not have a monitor_delay applied (since
the monitor has nothing to do with the data recorded from the
running wheel).