Factor out code to get change_frame for a trial
This is being done because there is indication from the VBN data
release that incrementing change_frame by 1 is incorrect (despite
the comment assuring us that MPE verified this procedure circa
October 2021). Changing this behavior outright will have implications
for the VBO data. We do not have time to resolve that conversation,
so we are factoring this logic out so that VBN can implement its
own Trial class that does not increment change frame by 1.

When SDK issue #1539 is resolved, we may decide that the appropriate
action is to undo this change (there will still need to be a specific
Trial child class for VBN for other reasons).